+++ packages/collections-common/src/CountStrategy.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Domain\Collections;
+namespace Rekalogika\Domain\Collections\Common;
 
 enum CountStrategy
 {