+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * build: initial commit
 * fix: rename `matching()` to `withCriteria()` to avoid conflict with PHPStan
   `allCollectionsSelectable`
+* refactor: split common code to `rekalogika/collections-common`
+* feat: new package `rekalogika/collections-orm`